Fixed typing of Object.groupBy
`Partial<Record<>>` is not a good enough type. It says each element may be undefined, but that's not true. Instead, the key used to index the object *might not exist at all*.

The difference in practice is that when using Object.values and Object.Entries on a result from Object.groupBy, `Partial<Record<>>` results in an array where each element can be undefined. In reality, the element would't make it into the array at all.